refacto + create test_utils
