+++ crates/client/db/tests/test_block.rs
@@ -18,6 +18,7 @@ use mp_transactions::{
     DeployAccountTransactionV3, DeployTransaction, InvokeTransactionV0, InvokeTransactionV1, InvokeTransactionV3,
     L1HandlerTransaction,
 };
+use mp_utils::tests_common::*;
 use rstest::*;
 use starknet_types_core::felt::Felt;
