+++ crates/client/db/tests/test_open.rs
@@ -3,6 +3,7 @@ mod common;
 use common::*;
 use mc_db::DatabaseService;
 use mp_chain_config::ChainConfig;
+use mp_utils::tests_common::*;
 use rstest::*;
 
 #[rstest]
