+++ crates/client/devnet/Cargo.toml
@@ -36,6 +36,7 @@ mp-convert.workspace = true
 mp-receipt.workspace = true
 mp-state-update.workspace = true
 mp-transactions.workspace = true
+mp-utils.workspace = true
 
 # Starknet
 blockifier.workspace = true
