+++ crates/client/devnet/src/lib.rs
@@ -219,8 +219,7 @@ mod tests {
         BroadcastedTransaction, DataAvailabilityMode, DeclareTransactionResult, DeployAccountTransactionResult,
         FlattenedSierraClass, InvokeTransactionResult, ResourceBounds, ResourceBoundsMapping,
     };
-    use std::env;
-    use std::path::PathBuf;
+    use mp_utils::tests_common::*;
     use std::sync::Arc;
 
     struct DevnetForTesting {
@@ -350,21 +349,6 @@ mod tests {
         .unwrap();
 
         DevnetForTesting { backend, contracts, block_production, mempool }
-    }
-
-    #[fixture]
-    pub fn set_workdir() {
-        let output = std::process::Command::new("cargo")
-            .arg("locate-project")
-            .arg("--workspace")
-            .arg("--message-format=plain")
-            .output()
-            .expect("Failed to execute command");
-
-        let cargo_toml_path = String::from_utf8(output.stdout).expect("Invalid UTF-8");
-        let project_root = PathBuf::from(cargo_toml_path.trim()).parent().unwrap().to_path_buf();
-
-        env::set_current_dir(&project_root).expect("Failed to set working directory");
     }
 
     #[rstest]
