+++ crates/client/eth/src/l1_messaging.rs
@@ -231,6 +231,7 @@ mod l1_messaging_tests {
     use mc_db::DatabaseService;
     use mc_metrics::MetricsService;
     use mp_chain_config::ChainConfig;
+    use mp_utils::tests_common::*;
     use rstest::*;
     use starknet_api::core::Nonce;
     use tempfile::TempDir;
@@ -314,21 +315,6 @@ mod l1_messaging_tests {
             }
         }
     );
-
-    #[fixture]
-    pub fn set_workdir() {
-        let output = std::process::Command::new("cargo")
-            .arg("locate-project")
-            .arg("--workspace")
-            .arg("--message-format=plain")
-            .output()
-            .expect("Failed to execute command");
-
-        let cargo_toml_path = String::from_utf8(output.stdout).expect("Invalid UTF-8");
-        let project_root = PathBuf::from(cargo_toml_path.trim()).parent().unwrap().to_path_buf();
-
-        env::set_current_dir(&project_root).expect("Failed to set working directory");
-    }
 
     /// Common setup for tests
     ///
