+++ crates/client/rpc/Cargo.toml
@@ -31,6 +31,7 @@ mp-convert = { workspace = true, default-features = true }
 mp-receipt = { workspace = true }
 mp-state-update = { workspace = true }
 mp-transactions = { workspace = true }
+mp-utils = { workspace = true }
 
 # Starknet
 blockifier = { workspace = true, default-features = true }
