+++ crates/client/rpc/src/test_utils.rs
@@ -14,12 +14,13 @@ use mp_state_update::{
     StorageEntry,
 };
 use mp_transactions::{InvokeTransaction, InvokeTransactionV0, Transaction};
+use mp_utils::tests_common::*;
 use rstest::fixture;
 use starknet_core::types::{
     BroadcastedDeclareTransaction, BroadcastedDeployAccountTransaction, BroadcastedInvokeTransaction,
     DeclareTransactionResult, DeployAccountTransactionResult, Felt, InvokeTransactionResult,
 };
-use std::{env, path::PathBuf, sync::Arc};
+use std::sync::Arc;
 
 use crate::{providers::AddTransactionProvider, Starknet};
 
@@ -47,21 +48,6 @@ impl AddTransactionProvider for TestTransactionProvider {
     ) -> RpcResult<InvokeTransactionResult> {
         unimplemented!()
     }
-}
-
-#[fixture]
-pub fn set_workdir() {
-    let output = std::process::Command::new("cargo")
-        .arg("locate-project")
-        .arg("--workspace")
-        .arg("--message-format=plain")
-        .output()
-        .expect("Failed to execute command");
-
-    let cargo_toml_path = String::from_utf8(output.stdout).expect("Invalid UTF-8");
-    let project_root = PathBuf::from(cargo_toml_path.trim()).parent().unwrap().to_path_buf();
-
-    env::set_current_dir(&project_root).expect("Failed to set working directory");
 }
 
 #[fixture]
