+++ crates/primitives/chain_config/Cargo.toml
@@ -14,6 +14,10 @@ blockifier = { workspace = true }
 starknet-types-core = { workspace = true }
 starknet_api = { workspace = true }
 
+# Madara
+
+mp-utils.workspace = true
+
 # Other
 anyhow = { workspace = true }
 lazy_static = { workspace = true }
