+++ crates/primitives/chain_config/src/chain_config.rs
@@ -235,29 +235,13 @@ impl ChainConfig {
 
 #[cfg(test)]
 mod tests {
-    use std::env;
-
     use blockifier::{transaction::transaction_types::TransactionType, versioned_constants::ResourceCost};
+    use mp_utils::tests_common::*;
     use rstest::*;
     use serde_json::Value;
     use starknet_types_core::felt::Felt;
 
     use super::*;
-
-    #[fixture]
-    fn set_workdir() {
-        let output = std::process::Command::new("cargo")
-            .arg("locate-project")
-            .arg("--workspace")
-            .arg("--message-format=plain")
-            .output()
-            .expect("Failed to execute command");
-
-        let cargo_toml_path = String::from_utf8(output.stdout).expect("Invalid UTF-8");
-        let project_root = PathBuf::from(cargo_toml_path.trim()).parent().unwrap().to_path_buf();
-
-        env::set_current_dir(&project_root).expect("Failed to set working directory");
-    }
 
     #[rstest]
     fn test_mainnet_from_yaml(_set_workdir: ()) {
