+++ crates/tests/src/lib.rs
@@ -3,10 +3,10 @@
 mod rpc;
 
 use anyhow::bail;
-use rstest::{fixture, rstest};
+use mp_utils::tests_common::*;
+use rstest::rstest;
 use starknet_providers::Provider;
 use starknet_providers::{jsonrpc::HttpTransport, JsonRpcClient, Url};
-use std::env;
 use std::ops::Range;
 use std::sync::Mutex;
 use std::{
@@ -200,21 +200,6 @@ impl MadaraCmdBuilder {
             _port: self.port,
         }
     }
-}
-
-#[fixture]
-pub fn set_workdir() {
-    let output = std::process::Command::new("cargo")
-        .arg("locate-project")
-        .arg("--workspace")
-        .arg("--message-format=plain")
-        .output()
-        .expect("Failed to execute command");
-
-    let cargo_toml_path = String::from_utf8(output.stdout).expect("Invalid UTF-8");
-    let project_root = PathBuf::from(cargo_toml_path.trim()).parent().unwrap().to_path_buf();
-
-    env::set_current_dir(&project_root).expect("Failed to set working directory");
 }
 
 #[rstest]
